% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSE.R
\docType{package}
\name{LSE}
\alias{LSE}
\title{LSE package}
\description{
Solve the equality constrained least sqaures problem for real value and for quaternions, also allows to use the Generalized QR factorization for two matrices simlutaneously.
}
\author{
Sergio Andrés Cabrera Miranda \email{sergio05acm@gmail.com}
}
